robokit-monorepo/pyproject.toml
[project]
name = "robokit-monorepo"
version = "0.1.0"
description = "RoboKit Monorepo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "typer>=0.16.0",
]
